Refactor form to own component, change formState to provide/inject
